# One simple SIMD optimization