--- a/src/unicode.cpp
+++ b/src/unicode.cpp
@@ -10,7 +10,10 @@ ADA_POP_DISABLE_WARNINGS
 #include <algorithm>
 #if ADA_NEON
 #include <arm_neon.h>
+#elif ADA_SSE2
+#include <emmintrin.h>
 #endif
+
 namespace ada::unicode {
 
 constexpr bool to_lower_ascii(char* input, size_t length) noexcept {
@@ -45,24 +48,54 @@ constexpr bool to_lower_ascii(char* input, size_t length) noexcept {
 ada_really_inline bool has_tabs_or_newline(
     std::string_view user_input) noexcept {
   size_t i = 0;
-  uint8x16_t mask1 = vmovq_n_u8('\r');
-  uint8x16_t mask2 = vmovq_n_u8('\n');
-  uint8x16_t mask3 = vmovq_n_u8('\t');
+  const uint8x16_t mask1 = vmovq_n_u8('\r');
+  const uint8x16_t mask2 = vmovq_n_u8('\n');
+  const uint8x16_t mask3 = vmovq_n_u8('\t');
   uint8x16_t running{0};
   for (; i + 15 < user_input.size(); i += 16) {
     uint8x16_t word = vld1q_u8((const uint8_t*)user_input.data() + i);
-    running = vorrq_u8(vorrq_u8(running, vorrq_u8(vceqq_u8(word, mask1),vceqq_u8(word, mask2))), vceqq_u8(word, mask3));
+    running = vorrq_u8(vorrq_u8(running, vorrq_u8(vceqq_u8(word, mask1),
+                                                  vceqq_u8(word, mask2))),
+                       vceqq_u8(word, mask3));
   }
   if (i < user_input.size()) {
     uint8_t buffer[16];
     memcpy(buffer, user_input.data() + i, user_input.size() - i);
     uint8x16_t word = vld1q_u8((const uint8_t*)user_input.data() + i);
-    running = vorrq_u8(vorrq_u8(running, vorrq_u8(vceqq_u8(word, mask1),vceqq_u8(word, mask2))), vceqq_u8(word, mask3));
+    running = vorrq_u8(vorrq_u8(running, vorrq_u8(vceqq_u8(word, mask1),
+                                                  vceqq_u8(word, mask2))),
+                       vceqq_u8(word, mask3));
   }
   return vmaxvq_u8(running) != 0;
 }
+#elif ADA_SSE2
+ada_really_inline bool has_tabs_or_newline(
+    std::string_view user_input) noexcept {
+  size_t i = 0;
+  const __m128i mask1 = _mm_set1_epi8('\r');
+  const __m128i mask2 = _mm_set1_epi8('\n');
+  const __m128i mask3 = _mm_set1_epi8('\t');
+  __m128i running{0};
+  for (; i + 15 < user_input.size(); i += 16) {
+    __m128i word = _mm_loadu_si128((const __m128i*)(user_input.data() + i));
+    running = _mm_or_si128(
+        _mm_or_si128(running, _mm_or_si128(_mm_cmpeq_epi8(word, mask1),
+                                           _mm_cmpeq_epi8(word, mask2))),
+        _mm_cmpeq_epi8(word, mask3));
+  }
+  if (i < user_input.size()) {
+    uint8_t buffer[16];
+    memcpy(buffer, user_input.data() + i, user_input.size() - i);
+    __m128i word = _mm_loadu_si128((const __m128i*)buffer);
+    running = _mm_or_si128(
+        _mm_or_si128(running, _mm_or_si128(_mm_cmpeq_epi8(word, mask1),
+                                           _mm_cmpeq_epi8(word, mask2))),
+        _mm_cmpeq_epi8(word, mask3));
+  }
+  return _mm_movemask_epi8(running) != 0;
+}
 #else
-ada_really_inline constexpr bool has_tabs_or_newline(
+ada_really_inline bool has_tabs_or_newline(
     std::string_view user_input) noexcept {
   auto has_zero_byte = [](uint64_t v) {
     return ((v - 0x0101010101010101) & ~(v)&0x8080808080808080);